# perf(table): core table performance improvements