--- a/packages/@ourworldindata/types/src/domainTypes/CoreTableTypes.ts
+++ b/packages/@ourworldindata/types/src/domainTypes/CoreTableTypes.ts
@@ -50,6 +50,7 @@ export enum TransformType {
     LoadFromColumnStore = "LoadFromColumnStore",
     LoadFromMatrix = "LoadFromMatrix",
     Concat = "Concat",
+    Noop = "Noop", // for when we notice that we can skip a transform
 
     // Row ops
     FilterRows = "FilterRows",