I found a good bunch of optimizations that we can apply to make core table methods more performant, especially for the our worst offender charts.

| Method                               | URL                                                                                   | Before: longest single call | Before: overall time | After: longest single call | After: overall time |
|--------------------------------------|---------------------------------------------------------------------------------------|-----------------------------|----------------------|----------------------------|---------------------|
| rowIndex                             | [covid-19-cumulative-confirmed-cases-vs-confirmed-deaths](http://localhost:3030/grapher/covid-19-cumulative-confirmed-cases-vs-confirmed-deaths) | 214ms                       | 501ms                | 131ms                      | 355ms               |
| sortBy                               | [covid-19-cumulative-confirmed-cases-vs-confirmed-deaths](http://localhost:3030/grapher/covid-19-cumulative-confirmed-cases-vs-confirmed-deaths) | 124ms                       | 296ms                | 100ms                      | 224ms               |
| interpolateColumnsByClosestTimeMatch | [covid-19-cumulative-confirmed-cases-vs-confirmed-deaths](http://localhost:3030/grapher/covid-19-cumulative-confirmed-cases-vs-confirmed-deaths) |                             | 763ms                |                            | 338ms               |
|                                      | [covid-19-daily-tests-vs-daily-new-confirmed-cases](http://localhost:3030/grapher/covid-19-daily-tests-vs-daily-new-confirmed-cases)       |                             | 350ms                |                            | 190ms               |
|                                      | [growth-of-income-and-trade](http://localhost:3030/grapher/growth-of-income-and-trade)                              |                             | 136ms                |                            | 80ms                |


The SVG tester duration also goes down (measured with `--isolate`):
- For the top-50 heaviest charts (many of them are scatters), down from 150s to 130s (-15%)
- overall runtime goes down from 988s to 941s (-5%)